Hill Climb takes in population and fitness test
Hill climb will take in an initial population of OurCreatures.  After
the first one, it will use the mutated populations to do future hill
climbs on.  Fitness test simulates a jump, however, due to an issue in
getFitness, jumps are useless because the jump is not usually less than
the current fitness the creature starts with.