Hill Climbing constructor and getBestFitDNA updated and trying to get fitness up
Hill Climbing now takes in a ArrayList<DNA> instead of OurCreature to
have creatures only being made in the fitness test and are immediately
detached.  Changed getCreature to getBestFitDNA.  updateBrain seems
kind of wonky.  Needs all the time that has passed to get a non 0
fitness, but physics don’t seem to be applying the the creature because
they are getting really high fitnesses, but in the simulation they are
no where near that high