HC keeps track of fitness deviation and method for signal to GA
HC will now keep track of the deviation of the fitnesses between the
current generation and the previous one.  If fitness is worse 3 times,
then mutation is needed, or if the abs deviation of the current fitness
and previous fitness is < the stall deviation, 1 right now, then
mutation is needed.