Hill Climbing Adjusts Joints For Mutated Block Size
Hill climbing now can adjust the joints for blocks that are having
their size being mutated, as well as, fixing the joint positions of
other blocks that have ids greater than the mutated block id.  Also
have in place in MainSim for GA to start when Hill Climbing needs a
boost.